PRE-RELEASE 1
Moved the "resting" dialogue for post-analysis out of the conditionals for searching
for a ScienceSpot and consoles. The resting dialogue will now simply occupy the space
ABOVE everything else EVEN if a cooldown is in progress.

Edited the part.cfg to reference Squad model to prevent directly distributing model
parts.

Everything readied for pre-release 1. Now to playtesting.