Work to PR3 - ii
Issue #19 - more accurate representation of "TURN LEFT" and "TURN RIGHT"
implemented. Before heading > spotbearing means to turn left.
A problem comes up if the spot bearing is 350 and heading is 010. Which is
not heading > spotbearing.

The workaround is to shift the science bearing to 0, and apply that shift
to the rover heading. That way if the rover heading is less than 180, turn left; vice-versa.

New values for science:
1% 400-500
8% 200-400
45% 70-200
70% 30-70
else 0-30