Class RSUpgrades added.
Completed RSUpgrades, which holds two methods that getCost and getValue for a specified enum (distanceLevel or maxAccuracyLevel), the values of which are stored in a class within RSUpgrades called "levels".

Now onto make the upgrade menu, then implementing the upgrades themselves.

Predicting accuracy method will be in RoverScience methinks.

Now construction RoverScienceGUI via RoverScience's OnAwake and passing upgrades
through. Hopefully it stays intact - haven't tried this kind before so
I guess we're about to see. MAYBE upgrades should be called as a static variable in
RoverScience - that should be okay.