Removed RSUpgrades class
RSUpgrades seemed to be breaking the plugin. Too many problems and too much workaround - defeats the purpose
of simplicity the class was meant to do.

This means levels will now be stored in RoverScience and so will the enum and the methods to
extract cost/values.

Save/load has now been re-implemented and is (should be) working fine.